+++ src/containers/ListingPage/SectionAuthorMaybe.js
@@ -1,6 +1,10 @@
 import React from 'react';
 import { FormattedMessage } from '../../util/reactIntl';
-import { INQUIRY_PROCESS_NAME, resolveLatestProcessName } from '../../transactions/transaction';
+import {
+  isInquiryProcessAlias,
+  isNegotiationProcessAlias,
+  OFFER,
+} from '../../transactions/transaction';
 
 import { Heading, Modal } from '../../components';
 import UserCard from './UserCard/UserCard';
@@ -30,8 +34,10 @@ const SectionAuthorMaybe = props => {
   }
 
   const transactionProcessAlias = listing?.attributes?.publicData?.transactionProcessAlias || '';
-  const processName = resolveLatestProcessName(transactionProcessAlias.split('/')[0]);
-  const isInquiryProcess = processName === INQUIRY_PROCESS_NAME;
+  const unitType = listing?.attributes?.publicData?.unitType || '';
+  const isInquiryProcess = isInquiryProcessAlias(transactionProcessAlias);
+  const isNegotiationProcess = isNegotiationProcessAlias(transactionProcessAlias);
+  const showContact = !(isInquiryProcess || (isNegotiationProcess && unitType === OFFER));
 
   return (
     <section id="author" className={css.sectionAuthor}>
@@ -42,7 +48,7 @@ const SectionAuthorMaybe = props => {
         user={listing.author}
         currentUser={currentUser}
         onContactUser={onContactUser}
-        showContact={!isInquiryProcess}
+        showContact={showContact}
         contactLinkId={CONTACT_USER_LINK}
       />
       <Modal