+++ CHANGELOG.md
@@ -14,6 +14,8 @@ way to update this template, but currently, we follow a pattern:
 
 ## Upcoming version 2026-XX-XX
 
+- [fix] ListingPage.duck.js: fix a bug where inquiry is not allowed for customer role on OFFER flow.
+  [#768](https://github.com/sharetribe/web-template/pull/768)
 - [fix] EditListingPage.duck.js: fix a bug where exception cache was not updated correctly.
   [#766](https://github.com/sharetribe/web-template/pull/766)